+++ assets/javascripts/kitten/karl/lazy/lazy-hook.js
@@ -1,0 +1,11 @@
+import React, { useEffect, useRef, useState } from 'react'
+import useLazyObserver from '../../helpers/utils/lazy-hook'
+
+const LazyHook = ({ children }) => {
+  const lazyComponentRef = useRef(null)
+  const isVisible = useLazyObserver(lazyComponentRef)
+
+  return <div ref={lazyComponentRef}>{isVisible && children}</div>
+}
+
+export default LazyHook