+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: new `getColor` helper with more specific color names
+
 ## [2.11.0] - 2019-06-20
 
 Features: