+++ CHANGELOG.md
@@ -6,6 +6,17 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Feature: Create `Autocomplete` component.
 
+## [2.11.0] - 2019-06-20
+
+Features:
+- Add `imageContainerBackground` and `imageContainerRatio` props to
+  `SimpleCard`.
+- Add Lazy class `LazyObserver`, hook `useLazyObserver` and HOC `withLazy`.
+
+Fixes:
+- `SimpleCarousel` accessibility fixes.
+- Remove warnings in `Carousel` component tests.
+
 ## [2.10.0] - 2019-06-07
 
 Feature: